Moves accounts-db stats into a submodule
